claude-computer-use-proxy/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-computer-use-proxy"
version = "0.1.0"
description = "Proxy-ready, Windows-native Claude-style computer use without Docker."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

